chore: move repositories to dedicated namespace
